health_whisperer_streamlit/requirements.txt
streamlit>=1.36
supabase==2.5.1
postgrest==0.15.3
httpx==0.27.2
httpcore==1.0.5
python-telegram-bot==21.4
google-generativeai>=0.7.0
pytz==2024.1
matplotlib